Cardano.Api.Any

Documentation

data Bech32DecodeError

class (HasTypeProxy a, SerialiseAsRawBytes a) ⇒ SerialiseAsBech32 a where

Methods

bech32PrefixFor ∷ a → Text

The human readable prefix to use when encoding this value to Bech32.

bech32PrefixesPermitted ∷ AsType a → [Text]

The set of human readable prefixes that can be used for this type.

data FileError e

Constructors

FileError FilePath e

FileErrorTempFile

Fields

FilePath

Target path

FilePath

Temporary path

Handle

FileIOError FilePath IOException

data ErrorAsException where

Constructors

ErrorAsException ∷ Error e ⇒ e → ErrorAsException

class Show e ⇒ Error e where

Methods

displayError ∷ e → String

class HasTypeProxy a ⇒ SerialiseAsCBOR a where

Minimal complete definition

Nothing

Methods

serialiseToCBOR ∷ a → ByteString

default serialiseToCBOR ∷ ToCBOR a ⇒ a → ByteString

deserialiseFromCBOR ∷ AsType a → ByteString → Either DecoderError a

default deserialiseFromCBOR ∷ FromCBOR a ⇒ AsType a → ByteString → Either DecoderError a

class HasTypeProxy a ⇒ SerialiseAsRawBytes a where

Methods

serialiseToRawBytes ∷ a → ByteString

deserialiseFromRawBytes ∷ AsType a → ByteString → Maybe a

class CastHash roleA roleB where

Methods

castHash ∷ Hash roleA → Hash roleB

data family Hash keyrole ∷ Type

data FromSomeType (c ∷ Type → Constraint) b where

Constructors

FromSomeType ∷ c a ⇒ AsType a → (a → b) → FromSomeType c b

class HasTypeProxy t where

Associated Types

data AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

Methods

proxyToAsType ∷ Proxy t → AsType t

serialiseToRawBytesHex ∷ SerialiseAsRawBytes a ⇒ a → ByteString

serialiseToRawBytesHexText ∷ SerialiseAsRawBytes a ⇒ a → Text

deserialiseFromRawBytesHex ∷ SerialiseAsRawBytes a ⇒ AsType a → ByteString → Maybe a

throwErrorAsException ∷ Error e ⇒ e → IO a

The preferred approach is to use Except or ExceptT, but you can if necessary use IO exceptions.

serialiseToBech32 ∷ a → Text

deserialiseFromBech32 ∷ AsType a → Text → Either Bech32DecodeError a

class Typeable a ⇒ FromCBOR a where

Minimal complete definition

fromCBOR

Methods

fromCBOR ∷ Decoder s a

label ∷ Proxy a → Text

class Typeable a ⇒ ToCBOR a where

Minimal complete definition

toCBOR

Methods

toCBOR ∷ a → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy a → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [a] → Size

data Proxy (t ∷ k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy