Cardano.Api.Key

Documentation

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.

class CastSigningKeyRole keyroleA keyroleB where

Methods

castSigningKey ∷ SigningKey keyroleA → SigningKey keyroleB

Change the role of a SigningKey, if the representation permits.

class CastVerificationKeyRole keyroleA keyroleB where

Some key roles share the same representation and it is sometimes legitimate to change the role of a key.

Methods

castVerificationKey ∷ VerificationKey keyroleA → VerificationKey keyroleB

Change the role of a VerificationKey, if the representation permits.

class (Eq (VerificationKey keyrole), Show (VerificationKey keyrole), SerialiseAsRawBytes (Hash keyrole), HasTextEnvelope (VerificationKey keyrole), HasTextEnvelope (SigningKey keyrole)) ⇒ Key keyrole where

An interface for cryptographic keys used for signatures with a SigningKey and a VerificationKey key.

This interface does not provide actual signing or verifying functions since this API is concerned with the management of keys: generating and serialising.

Associated Types

data VerificationKey keyrole ∷ Type

The type of cryptographic verification key, for each key role.

data SigningKey keyrole ∷ Type

The type of cryptographic signing key, for each key role.

Methods

getVerificationKey ∷ SigningKey keyrole → VerificationKey keyrole

Get the corresponding verification key from a signing key.

deterministicSigningKey ∷ AsType keyrole → Seed → SigningKey keyrole

Generate a SigningKey deterministically, given a Seed. The required size of the seed is given by deterministicSigningKeySeedSize.

deterministicSigningKeySeedSize ∷ AsType keyrole → Word

verificationKeyHash ∷ VerificationKey keyrole → Hash keyrole

generateSigningKey ∷ Key keyrole ⇒ AsType keyrole → IO (SigningKey keyrole)

Generate a SigningKey using a seed from operating system entropy.