Cardano.Api.OperationalCertificate

Description

Operational certificates

Documentation

data OperationalCertificate

Constructors

OperationalCertificate !(OCert StandardCrypto) !(VerificationKey StakePoolKey)

data OperationalCertificateIssueCounter

Constructors

OperationalCertificateIssueCounter

Fields

opCertIssueCount ∷ !Word64

opCertIssueColdKey ∷ !(VerificationKey StakePoolKey)

newtype KESPeriod

Constructors

KESPeriod

Fields

unKESPeriod ∷ Word

getHotKey ∷ OperationalCertificate → VerificationKey KesKey

getKesPeriod ∷ OperationalCertificate → Word

getOpCertCount ∷ OperationalCertificate → Word64

Data family instances

data family AsType t

A family of singleton types used in this API to indicate which type to use where it would otherwise be ambiguous or merely unclear.

Values of this type are passed to deserialisation functions for example.