Cardano.Tools.DBAnalyser.Analysis.BenchmarkLedgerOps.SlotDataPoint

Documentation

data SlotDataPoint

Information about the time spent processing the block corresponding to slot, divided into the five major operations:

Forecast.

Header tick.

Header application.

Block tick.

Block application.

It is up to the user of a slot data point to decide which units the data represent (eg milliseconds, nanoseconds, etc)

Constructors

SlotDataPoint

Fields

slot ∷ !SlotNo

Slot in which the 5 ledger operations were applied.

slotGap ∷ !Word64

Gap to the previous slot.

totalTime ∷ !Int64

Total time spent in the 5 ledger operations at slot.

mut ∷ !Int64

Time spent by the mutator while performing the 5 ledger operations at slot.

gc ∷ !Int64

Time spent in garbage collection while performing the 5 ledger operations at slot.

majGcCount ∷ !Word32

Total number of major garbage collections that took place while performing the 5 ledger operations at slot.

mut_forecast ∷ !Int64

Difference of the GC.mutator_elapsed_ns field when computing the forecast.

mut_headerTick ∷ !Int64

mut_headerApply ∷ !Int64

mut_blockTick ∷ !Int64

mut_blockApply ∷ !Int64

showData ∷ SlotDataPoint → Builder → Builder

showHeaders ∷ Builder → Builder

Return the headers that correspond to the fields of SlotDataPoint.

The position of each header matches the position in which the corresponding field value is returned in showData. Eg, if show headers returns:

"slot slotGap totalTime" ...

then the third value returned by showData will correspond to totalTime.